requirements.txt
aiogram==2.25.2
python-dotenv==1.0.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
